fix warnings

+++ arduino/Thermometer/Thermometer.ino
@@ -7,8 +7,9 @@ BLEDescriptor temperatureDescriptor = BLEDescriptor("2901", "degrees C");
 
 #define TEMPERATURE_PIN A0
 
-long previousMillis = 0;  // will store last time temperature was updated
-long interval = 2000;     // interval at which to read temperature (milliseconds)
+unsigned long previousMillis = 0;  // will store last time temperature was updated
+unsigned short interval = 2000;    // interval at which to read temperature (milliseconds)
+
 void setup()
 {
   Serial.begin(9600);
